added a check-up before deleting the product

+++ backend/models.py
@@ -54,11 +54,25 @@ def update_product(product_id, product_data):
     finally:
         if conn:
             conn.close()
+
 def delete_product(product_id):
+    conn = None
     try:
         conn = sqlite3.connect('./database/supermarket.db')
         cursor = conn.cursor()
 
+        # перевірка на те, чи існує продукт у таблиці product
+        cursor.execute('''
+            SELECT id_product FROM product WHERE id_product = ?
+        ''', (product_id,))
+        product_exists = cursor.fetchone()
+
+        if not product_exists:
+            return {
+                "status_code": 404,
+                "body": jsonify({"data": "Product not found"}),
+                "headers": {"Content-Type": "application/json"}
+            }
         cursor.execute('''
             DELETE FROM store_product
             WHERE id_product = ?
@@ -76,6 +90,7 @@ def delete_product(product_id):
             "body": jsonify({"data": "Product deleted successfully"}),
             "headers": {"Content-Type": "application/json"}
         }
+
     except sqlite3.Error as e:
         return {
             "status_code": 500,
